Update main.tf
I'm on the Azure AI services documentation team. According to PM, we recommend that customers start using the AIServices "kind" instead of the "CognitiveServices" kind. So, we want the sample to show how to create AIServices.

This PR might take a dependency on one of these PRs to complete first.

https://github.com/hashicorp/terraform-provider-azurerm/pull/26713
 
https://github.com/hashicorp/terraform-provider-azurerm/pull/26008